broke alignment but victory over card placement

+++ core/src/com/pear/pudding/model/Board.java
@@ -21,7 +21,7 @@ public class Board extends Deck {
         setLocation(Location.BOARD);
         setSlotWidth(width / NUMBER_OF_BOARD_SLOTS);
 
-        float middleSlotIndex = NUMBER_OF_BOARD_SLOTS / 2;
+        int middleSlotIndex = NUMBER_OF_BOARD_SLOTS / 2;
         for (int i = 0; i < NUMBER_OF_BOARD_SLOTS; i++) {
             Side slotSide = i < middleSlotIndex ? Side.LEFT : (i > middleSlotIndex ? Side.RIGHT : Side.MIDDLE);
             var slot = new Slot(getX() + i * getSlotWidth(), getY(), getSlotWidth(), getHeight(), i);
